move properties things from App server to a proper lib (in the wrong folder) made a new method to get property by key.
there is a new file in the root of project (conf) that have the user's configurations,
this file can be changed.
