+++ conf
@@ -1,0 +1,2 @@
+#raspberryGameSuiteServer user properties
+hostname = 192.168.1.108
